Moved buttonListener Checks into file
Created utill file for timeExemptCheck and timeExemptMessage and made them into files returning bool values

CodeBase is now that much cleaner
